cpu backend refactor and unit tests pass.
