+++ src/ngraph/CMakeLists.txt
@@ -130,6 +130,8 @@ set (SRC
     op/erf.hpp
     op/exp.cpp
     op/exp.hpp
+    op/experimental/batch_dot.cpp
+    op/experimental/batch_dot.hpp
     op/experimental/dyn_broadcast.cpp
     op/experimental/dyn_broadcast.hpp
     op/experimental/dyn_pad.cpp
