+++ src/ngraph/runtime/cpu/pass/cpu_mat_fusion.cpp
@@ -32,7 +32,7 @@
 #include "ngraph/op/slice.hpp"
 #include "ngraph/pattern/matcher.hpp"
 #include "ngraph/pattern/op/label.hpp"
-#include "ngraph/runtime/cpu/op/batch_dot.hpp"
+#include "ngraph/op/experimental/batch_dot.hpp"
 #include "ngraph/runtime/cpu/op/group_conv.hpp"
 #include "ngraph/util.hpp"
